sam_browser_onnx/requirements.txt
# SAM Browser ONNX App Requirements

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Deep learning
torch==2.1.0
torchvision==0.16.0
onnx==1.15.0

# SAM model
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Image processing
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3

# Optional: for packaging
# pyinstaller==6.2.0
